Change world news page latest section to use Rummager
This changes the latest section on World Location News pages to use Rummager to
retrieve its documents instead of the Whitehall database when the locale is set
to English. This allows content published by Content Publisher to surface in
this section. Currently, Content Publisher does not support translated content
so we ensure that the latest section continues to retrieve documents from the
Whitehall database when the locale is not English.

One side-effect of switching to use Rummager is that Rummager currently has no
concept of a published date vs. an updated date. Instead it provides a
`public_timestamp` field, so we now render this in the latest section instead
of "updated [timestamp]" and "published [timestamp]". This is consistent with
how dates are rendered in the announcements and publications sections, as well
as the latest section on topical events pages.