Change world news page announcements to use Rummager
This changes the announcements section on World Location News pages to use
Rummager to retrieve its documents instead of the Whitehall database when the
locale is set to English. This allows content published by Content Publisher to
surface in this section. Currently, Content Publisher does not support
translated content so we ensure that the announcements section continues to
retrieve documents from the Whitehall database when the locale is not English.