Change world location news atom feeds to use Rummager
This changes the World Location News atoms feed to use Rummager to retrieve its
documents instead of the Whitehall database when the locale is set to English.
This allows content published by Content Publisher to surface in the feed.
Currently, Content Publisher does not support translated content so we ensure
that the feed continues to retrieve documents from the Whitehall database when
the locale is not English.

There are a couple of user-facing changes as a result of this work:

1. The id of each entry in the feed will change because we are no longer
using the edition id, which only exists in the Whitehall database.

2. The entries will no longer display the body of the content as
Rummager doesn't store this information