Change reslug task to add edition to the search index
Previously, my thoughts were that sending the edition to the publishing
api would populate the govuk index via a message being enqued on the
rabbitmq que and rummager would pick the job up and index the document.

However, this did not happen when i last ran this rake task and i had to
republish the document for it to be indexed. This change manually adds
the edition to the index.