Add offsite link types for Content Publisher documents
This adds offsite link types for the two document types that Content Publisher
allows for publishing. This is intended as a short time fix (I know, I
know I wince writing those words) to allow featuring of Content
Publisher documents through Whitehall while the initial private beta is
in progress.

This adds the ability for the link type to be displayed in different
ways on the frontend and backend of Whitehall.