Move CIP publishing API guard
This guard is a workaround to prevent errors when saving corporate information pages to the publishing API, however it was only in place for non-translated CIPs.

Translated CIPs were triggered using a subclass of `EditionService` which did not implement the guard.

Moving the guard into the `update_publishing_api!` method makes it less likely to be bypassed by subclasses and the like.