Republish organisation after creating a contact
The links between organisations and contacts are added when the
organisation gets published. When creating new contacts, the
organisation doesn't get republished so the links aren't created
(even though the contact itself has made it to the content store).

I'm not sure why this has stopped working recently and how this
sort of dependency in Whitehall is meant to work so I've based
this code slightly on 9ab682152226dfc635848ac6e0a413e2208f2678.