Change the rendering_app for publish intents
To government-frontend. This is better than Whitehall frontend, as my
guess is that there isn't any content that can be scheduled through
Whitehall that's still rendered by whitehall-frontend.

This came out of investigating some missing Statistics Announcements,
and realising that there were lots of routes in the Router API with
the rendering-app as whitehall-frontend.