Fix visible_taxons to return all taxons
Instead of just the level one taxons. This fixes it's use in
TaxonomyTagForm, where it's used to remove taxons from the list of
invisible taxons.

Without this change, you can't untag content from level 2 or below
taxons, because they are treated as "invisible", even though they
appear on the form.

This has come up because of a Zendesk ticket.