Improve placement of cancel on prepare to close government
The placement of the cancel option on prepare to close government view
is very close to the "Yes, close this government" button. This could
easily lead to misclicks.

Update the form and move the cancel option
adjacent to the "Yes" button. This also makes it consistent with the way
Whitehall normally lays out its confirm and cancel actions on views.