Replace announcements lists with news and comms lists on world location news pages
We have created a new finder within finder-frontend
called _news and communications_.

This change ensures that the information under
news and communications on the /world/:slug/news
pages matches the content of the news and comms
finder (they both filter results based on
content_purpose_supergroup: news_and_communications).

The link to the announcements finder has also been
replaced.

The goal is to deprecate the announcements finder
and replace it with the news and communications finder;
replacing the links is required before we can do this.