Make it possible to redirect static pages
This will enable us to run a rake task to remove pages
published via PublishStaticPages such as
/government/announcements.

The new rake task redirect_static_pages will replace
specific static pages with redirects to new URLs and
will remove them from the search index.

The logic of RedirectStaticPages is a reverse of
PublishStaticPages, so if you read both that should
help explain the intent of the new class.