Add `publishing:scheduled:next_8_weeks` rake task
As Brexit approaches, we want to be able to know how many publications
are scheduled in the next few weeks. Dates are not hardcoded, so this
can be used in the future as well.

The result of this rake task is similar to this:
---
There are currently 606 scheduled publications in Whitehall.
Of all the scheduled publications:
 - 146 were scheduled before now (Tue, 12 Mar 2019 17:34:51 +0000)
 - 11 are scheduled between now (Tue, 12 Mar 2019 17:34:51 +0000) and 
the end of this week (Sun, 17 Mar 2019 23:59:59 +0000)
 - 20 are scheduled in the week beginning on Mon, 18 Mar 2019 00:00:00 
+0000
 - 24 are scheduled in the week beginning on Mon, 25 Mar 2019 00:00:00 
+0000
 - 9 are scheduled in the week beginning on Mon, 01 Apr 2019 00:00:00 
+0100
 - 11 are scheduled in the week beginning on Mon, 08 Apr 2019 00:00:00 
+0100
 - 14 are scheduled in the week beginning on Mon, 15 Apr 2019 00:00:00 
+0100
 - 9 are scheduled in the week beginning on Mon, 22 Apr 2019 00:00:00 
+0100
 - 5 are scheduled in the week beginning on Mon, 29 Apr 2019 00:00:00 
+0100
 - 11 are scheduled in the week beginning on Mon, 06 May 2019 00:00:00 
+0100
 - 346 are scheduled on or after Mon, 13 May 2019 00:00:00 +0100
---