Add `publishing:scheduled:at_same_time` rake task
As Brexit approaches, we want to be able to know how many publications
are scheduled to be published at the same time. There are no hardcoded
dates, so this can be used in the future as well.

The result of this rake task is similar to this:
---
In 157 instances, less than 5 publications are scheduled at the same
time.
In 43 instances, between 5 and 9 publications are scheduled at the same
time.
In 4 instances, 10 or more publications are scheduled at the same time.
The 4 instances in which 10 or more publications are scheduled at the
same time are divided as follows:
 - 25 publications scheduled on Thu, 28 Feb 2019 09:30:00 +0000
 - 15 publications scheduled on Thu, 21 Feb 2019 09:30:00 +0000
 - 15 publications scheduled on Tue, 19 Feb 2019 09:30:00 +0000
 - 12 publications scheduled on Fri, 29 Mar 2019 09:30:00 +0000
---