Improve loading attachments for publication
As part of our load testing improvements, we want to make the
publication overview page more resilient.

There are three types of `Attachments`: `FileAttachment`, `HtmlAttachment`,
and `ExternalAttachment`. They all share the same view when loaded
into the publication overview and currently calls the attachments of an
`edition` within the view. Each `attachment` requires it's `attachment_data`,
so including it here should optimise this query.

Co-authored-by: Alan Gabbianelli <alan.gabbianelli@digital.cabinet-office.gov.uk>