Add task to update in search editions published in a given range
The traffic replay from rummager to search-api doesn't seem to be 100%
reliable.  Over the weekend it looks like 7 documents didn't make it
across (a miss rate of about 5%), and resending them individually to
search fixed them.

Hopefully talking to search-api directly will be more reliable than
the traffic replay but, in any case, it looks like 2ndline will
occasionally need to update documents in search at least in the
immediate future.  So let's make it easier.

This is also a far easier alternative to reindexing Whitehall content
than doing the traffic replay.