Bundle video player in application javascript
This file was previously included by static however this was removed in
favour of govuk_publishing_components providing the functionality [1].

I had hoped to resolve the problem in this application by switching to
using govuk_publishing_components to wrap all frontend instances of
govspeak so that they could be initialised by
govuk_publishing_components. However this proved to be rather blocked by
govuk_publishing_components being significantly out of date for
Whitehall [2] and updating this being a non-trivial problem to fix.

This fix restores the video functionality across Whitehall frontend,
however this is achieved using the legacy approach to embedding videos
and barcharts in govspeak. so this may hit problems if/when Whitehall
frontend embraced the govspeak component.

[1]: https://github.com/alphagov/govuk_publishing_components/pull/751
[2]: https://github.com/alphagov/whitehall/pull/4720