Change SCSS compression approach
We currently can't update to the latest version of
govuk_publishing_components because of SCSS imported by govuk-frontend
of:

```
  @supports (margin: unquote("max(calc(0px))")) {
    $gutter-safe-area-right: calc(#{$govuk-gutter-half} + env(safe-area-inset-right));
    $gutter-safe-area-left: calc(#{$govuk-gutter-half} + env(safe-area-inset-left));

    // Use max() to pick largest margin, default or with safe area
    // Escaped due to Sass max() vs. CSS native max()
    margin-right: unquote("max(#{$govuk-gutter-half}, #{$gutter-safe-area-right})");
    margin-left: unquote("max(#{$govuk-gutter-half}, #{$gutter-safe-area-left})");
  }
```

This code does not work if a css_compressor is set to use :sass. The
reason this does not work is because the SCSS is effectively processed
twice. The first time, to generate the SCSS unquote is ran and a string
is unquoted. The second time, to compress the SCSS the unquoted string
is treated as a SCSS function and executed. The reason the
`max(calc(0px))` is quoted is because there is a SCSS & CSS naming
conflict on max. Thus the approach to resolve this is to remove the
css_compressor step and instead specify that the generation of SCSS
should be output compressed without line_comments.

For this to work we also have to remove the asset_bom_removal-rails gem
as that sets itself as a SCSS compressor and then falls back to the
default one. This gem was introduced to fix a problem in Firefox < 52
and since we are now on Firefox 66 this does not seem to be a concern.