Add length validation for featured_link url
The URL can have maximum of 255 characters as this is the default length of a
string in an ActiveRecord model. Users were getting a general “We're sorry, but
something went wrong” message when the URL they entered was too long.  This sets
the maxlength attribute on the edit form field and adds validation as well as
an error message in the model.