Add a rake task to add primary org to HTML publications.
Previously only `organisations` was being sent to the publishing-api
for html_publications.

We need `primary_publishing_organisation` to be in the payload we
receive in `content-data-api`, so we can filter on it.

This commit calls `publishing_api.patch_links` on each attachment from
all published documents, using the first of the `lead_organisations`
from the attachable.