Fix saving of HTML attachments
Previously we were attempting to get the primary organisation
from the first of the parent's `lead_organisations`

The broke saving of attachments on some document types that
don't implement `lead_organisations`.

This commit makes the code more defensive, getting the primary
org from `organisations` if `lead_organisations` is not implemented.