Add a Rake task for manually unpublishing content
Whitehall deletes some content when it unpublishes it, this makes it
very difficult to change the details of the unpublished
content (e.g. change from gone to redirect, or change the redirect).

This Rake task provides a workaround intended for these cases where
you want to manually send an unpublish request to the Publishing API.

This is motivated by Zendesk ticket 3694812.