Add a timeout on calculating the number of pages
Sometimes a malformed PDF will cause calculating the number of pages to
take a significant amount of time (we have one specific example that
takes over 2 minutes). This causes the user to see a timeout error when
uploading the PDF.

By adding a timeout, we can avoid the bad user experience of seeing a
timeout error, for the alternative of not having a number of pages set
for that attachment.

Ideally, we would restructure the way this code works to perform the
number of pages calculation in a background worker, however this would
require significant refactoring on how attachments work and this is a
pragmatic solution.