Adds redirects for publications finder atom feed
This adds in a check to see if an atom feed locale is set to English,
and if it is, redirects it towards a finder-frontend finder. The rules
for which finder are all set up in publications_routes.rb in the lib.

Also added a test for this, and stripped out old tests that revolved
around the atom feed while the locale is English - they're all
deprecated now!