Update org page public timestamp automatically
If we omit the public timestamp from the payload, then publishing-api will
generate one for us.  If we include it in the payload, then it should be
more accurate than this.

At present lots of things can change on an org page before the public
timestamp changes.  Before this change we were only updating the public
timestamp if a property of the organisation model had been updated.

Organisation pages include ministers, featured content, featured links,
other important members (such as high ranking civil servants or military
staff).  We really want the page to update when these things change,
particularly because it prompts external search engines to crawl them
again.  (Bing still seems to think that Gavin Williamson is Secretary
of State for Defence)

Organisation pages always do major updates on save, so it seems sensible
to update the public timestamp at the same time, whatever is edited.