Permit malformed array params when redirecting
When redirecting old finders publications/announcements
to the new finders served by finder-frontend, we don't
support malformed arrays in the query string. This fixes
that by fixing the malformed arrays before we redirect.

Example

?departments%5B0%5D=department-for-education
is parsed by rails as
{departments=>{0=>department-for-education}}
but really it should be
{departments=>[department-for-education]}

This uses the existing Whitehall::DocumentFilter::CleanedParams
class (which contains the method clean_malformed_array_params)
to fix this in a consistent manner.

Trello: https://trello.com/c/eTCe4Vun/825.