Use organisation_logo for WorldwideOrganisation
This comes from govuk_publishing_components so it will look consistent
with the rest of the site.

I've had to slightly modify the CSS for it to appear correctly on the
page.

Using this component does reduce the text size of the component, but I
think it's better to follow the design guide and stick with the size
specified by this component.