Fix topical events URL
Partly answers https://govuk.zendesk.com/agent/tickets/3732860

At present this is redirected to /search/all, but with a "topic" URL that
new search doesn't understand.

We need to add topical events to the list of things that the whitehall
search redirects will handle.

We'll address this more fully with planned work (by adding a facet etc.),
but this resolves the broken link issue for now.