Raise error if locked document is featured
If a document is locked and in the process of being migrated to Content
Publisher then we want to prevent it from being eligible to be featured so that
we can avoid broken links. This ensures that, at the controller level, an error
is raised if there is an attempt to feature a locked document.