Exclude locked documents from EditionFilter
The EditionFilter is used to generate a list of documents that are eligible to
be featured on organisation, topical event and world location pages. We want to
prevent locked documents from appearing in this list so that they aren't
accidentally featured, resulting in a broken link. To achieve this we ensure
that EditionFilter only returns a list of documents that are not locked.