Fix news article step
This fixes a failing test in the 'Email signup for people and roles' feature
test. It turns out that it has always been failing incorrectly; it looks for an
'email' link and clicks the incorrect one ('Don't have an email address?'),
meaning that the webdriver stays on the page instead of redirecting to the
correct email signup page, and incorrectly passes the assertion that the sign
up has actually occurred.

Recently a change to the component that renders this link meant that the 'Email
signup for people and roles' feature test started failing and it has at least
provided us with the opportunity to make it test the right thing. The Search
API returns a list of announcements related to a person and if there are any
then the People show page will display them along with subscription links,
including a link to subscribe to email alerts. In this test's case, the Search
API wasn't returning any announcements and as such the test failed to find an
'email' link. To fix this we have stubbed Search API to return a news article
that was associated to the Person described in the test.