Fix admin taxonomy helper assertion
An Admin Taxonomy Helper method has been failing as a result of an error
claiming that no arguments are being passed to the `assert_received` assertion.
This isn't actually the case, but it also turns out that this assertion is
provided by a deprecated gem, Bourne (https://github.com/thoughtbot/bourne). We
take the opportunity to get rid of this unsupported gem, since this assertion
is the only one provided by it, and replace it with `assert_respond_to`. It
does have the downside that we are unable to check that specific arguments have
been passed to the request however we're confident that the request itself is
being made and it feels like a reasonable thing to give up in order to get rid
of an unsupported gem.