Optionally include locked documents in EditionFilter
In a previous commit we prevented locked documents from being returned by the
EditionFilter to avoid them appearing in a list of documents that could be
featured. This adds a flag to include locked documents in the list returned by
EditionFilter as we want to display them on the index page.