Raise if PublishingApiWorker processes locked document
There are a number of workers that process edition changes and pass them on to
the Publishing API, as well as to other workers. We introduce a method to check
if a document is locked to the PublishingApiWorker, which other workers that
inherit from it can use. If the check fails, then an error is raised.

While PublishingApiServicesAndInformationWorker inherits from
PublishingApiWorker, it will not make use of this new method as it only deals
with organisations, which are unrelated to this check.