Add lang to untranslated office contacts
Some of our Worldwide Organisations have 'offices' or 'contacts'. Sometimes these are translated into the language the organisation page is in, but if not it will fallback to English.

This commit adds a function which detects if the fallback is being used (using the translated_locales function) and adds the lang attribute if necessary.