Add LockedDocumentConcern
We can now see that there are several places where we need to check for a
locked document outside of the Publishing API workers so it makes sense to
extract the PublishingApiWorker#check_if_locked_document method into a
LockedDocumentConcern that can be mixed in to other files across the app and
reduce repetition of the error. We anticipate adding further methods for places
where a content_id isn't available to look up a document.

We also add a custom LockedDocumentError to raise if a document is found to be
locked.