Add lang attribute to Corporate Information section
The links in the Corporate Information section always display as English, even on translated pages.
We therefore want to mark them up as English.
I've used a method to detect the translation rather than hardcoding it as English in case the functionality to do this is added in the future.