Disable Whitehall DB operations on a document when locked When a document has been migrated it is locked on Whitehall side. This commit takes steps to ensure that when locked, the user cannot edit the document at a database level even if UI functionality is exposed to allow it.
This commit:
- Adds a locking check on Feature, Document, Edition and
DocumentCollectionGroupMembership as a before_* step based on
the behaviour that could occur in that class (validation or update).
- Adds model tests for the above that ensures the expected behaviour.
- Modifies the way editions are created for existing document lock
tests such that they work with the new behaviour for edition lock checks.

This commit uses the new LockedDocumentConcern created for this purpose,
but relies on untidy logic to work around the issues with class vs
instance methods when passing the required objects to the concern.