Fix changing the timeout for the Publishing API client
Calling the publishing_api_with_low_timeout function would cause the
Publishing API client timeout to be set to 1, which would break parts
of Whitehall that talk to the Publishing API, casuing it to "hang up"
after a second.

Rather than using `.dup`, just make it a parameter to a new function,
and call that in both places.