Redirect association routes for locked documents
This ensures that if a user makes a request relating to an edition's various
associations, if the edition's document is locked then the user is redirected
to the document show page, preventing them from editing the document's
associations. Edition#forbid_editing_of_locked_documents has been extracted out
into the Admin:BaseController so that is can be reused by the other
controllers.