Redirect needs routes for locked documents
The Admin::NeedsController doesn't quite follow the same pattern as the other
controllers that are used for editing edition associations, insofar as it
doesn't perform a `find_edition` before hook, which means that an @edition
instance variable isn't set before the `forbid_editing_of_locked_documents`
hook is called, which it needs. This adds a `find_latest_edition` method to the
controller which allows us to set @edition for the locked documents hook as
well as for the update and edit actions. These actions also need @document set,
so we use `find_latest_edition` to set this as well.