Create index action in export/document_controller
This creates a endpoint which returns a json response of documents
information for a given lead_organisation and edition type.

lead_organisation is needed to find all the documents where the
lead_organisation passed in is present on the latest edition of each
document. `type` is used to specify the whitehall edition type of the
documents you want returning, e.g NewsArticle or Publication.

`lead_organisation` and `type` are mandatory params, `items_per_page`
and `page_number` are optional which will default to 100 and 1
respectively.