Display alternative show page for locked documents
We don't want users to be able to edit locked documents, so when they navigate
to a show page for a locked document we display a stripped-back version without
navigation to pages where the document could be edited. Instead we display only
key metadata, a notice explaining that the document has been moved to Content
Publisher, and a link to the document in Content Publisher.