Indicate locked documents on document index page
We add a label to locked documents on the document index page to indicate that
it has been moved to Content Publisher. Once the document has been moved it
won't make sense to display the "Updated by", "Updated" and "State" values as
it won't be in sync with what is in Content Publisher, so we replace these with
another message about having been moved to Content Publisher.