Add method to set document collection ordering by membership
This is to move away from document being the sole item in a document
collection group.

This loses the ability to add new items to the collection in this
context since a membership has to already exist. This can be worked
around though to maintain the same functionality for the user.