Report on non-accessible attachments by organisation
As part of meeting the new accessibility regulations, organisations
need to be more transparent on how accessible their content is on
GOV.UK. This includes reviewing their attachments that are not
accessible, and the steps they will take to improve them.

This rake task generates a CSV report containing information about
all published non-accessible attachments that belong to the specified
organisation. This should help organisations determine the amount
of work required.