Change document collection groups to work with memberships
This changes the Whitehall UI to work with both the document and
non-Whitehall link types of document collection memberships. This is
mostly achieved by switching the ids used in the UI and backend from
being documents to memberships.

The switch to this has required a change in the means memberships are
changed from group to group. Previously when working with documents this
could be changed with rather simple setting of documents to each group.
This can't be the case with memberships as if a membership is removed
from a group it is deleted. To resolve this memberships are moved to any
new groups first.

Finally, this removes a number of associations and methods that weren't
used or replaced with more appropriate options given their current
usage.