Create UI to add non-Whitehall document collection members
This adds the user interface component of adding a non-Whitehall link to
a document collection. In practice this is actually any link on GOV.UK.

The key parts of the logic for this are handled by a new model class
DocumentCollectionNonWhitehallLink::GovukUrl which handles validating
the user input, looking up data in Publishing API and persisting to the
database.

There are a few styling changes added to support this, the most notable
is the use of the details element to provide progressive disclosure. I
struggled to apply the same focus styling as similar elements with
mixins so ended up having to copy and paste some of it.