Use memberships as basis for collection ordering
Now that collections contain both documents and non-Whitehall links this
count is incorrect, this meant that items added were appearing in an
unexpected location. Switching to memberships covers both of these types
and corrects the issue.

With the switch to memberships there is now a complication that the item
may or may not be included in the memory of the
document_collection_group memberships depending on how the membership
was added. This is why this now has a conditional.