Setup redirects for old signup pages
For all urls that don't lead to finders these redirects take the user to
the equivalent signup page in email-alert-frontend.

For finders however this redirects the user to the publication,
announcement or statistics controllers with the query param in tact. This then
redirects the user to the new finders in finder frontend with the facets
pre-filled. This isn't an ideal approach but work would be needed in
email-alert-frontend to remove the step of taking a user to
finder-frontend first.

Example email signup url:
https://www.gov.uk/government/email-signup/new?email_signup%5Bfeed%5D=
https%3A%2F%2Fwww.gov.uk%2Fgovernment%2Fpublications.atom%3F
departments%255B%255D%3Deducation-and-skills-funding-agency%26publication_filter_option%3Dcorrespondence