Fix ordering approach for DocumentCollectionGroupMembership
This resolves a bug where copying a list of document collection group
memberships led to every item to having the same ordering (the maximum
value) which meant that all ordering was lost.

Instead this uses the index in the membership if it is available
otherwise it works out the maximum one.