Redirect html attachments to parent edition on deletion
Deleting an html attachment in Whitehall Admin sets it's deleted property to true and
prevents it from appearing in the admin panel for the appropriate edition.

Any previously published html attachments that are deleted on a subsequent edition,
 although no longer linked to, are still available at the original url.

This commit ensures that, when an edition is published, any deleted html attachments
are also redirected by the publishing-api to the parent editon.