Run organisations link checker task at 3am
We're seeing a problem where the link checker task means that Link
Checker API ends up trying to talk to Whitehall while the data sync is
in process, which means we see a lot of errors in Sentry.

By moving the task to 3am, the data sync will have finished by then, and
the link checking task itself doesn't take very long.