Change transaction callbacks to after_commit
On creating a new CorporateInformationPage, we need to republish the associated
Organisation to ensure that the new CIP is linked to.

The `after_save` callback causes the Organisation to be republished before the
database transaction has fully completed, and the CIP is still in `draft`, and
so the Organisation does not list it in it's list of published CIPs.

Changing this to `after_commit` ensures that the transaction has fully completed
before the Organisation is republished, and so the CIP will be `published`.

As `after_commit` also fires after a record is destroyed, we can remove the
`after_destroy` callback also.