Fix url whitespace bug
Previously if there was any whitespace in the URL for the related
content fields whitehall would crash throwing the standard 'Something
went wrong' message. This change strips all the whitespace from the URL
before parsing.

This problem became apparent when a zendesk came in which was related to
extra whitespace unknowingly added to the end of a URL.

Zendesk:
https://govuk.zendesk.com/agent/tickets/3831838