Translate news article type ID to descriptive name & value
IDs alone are meaningless and would require a lookup. It is
more useful to name the type than to refer to its ID.

Have had to serialise `output[:edition]` as JSON before doing
this transformation, as we've needed to write new fields, and
this would normally result in an ActiveRecord error along the
lines of "Unable to write new unexpected attribute". This was
because `edition` is an instance of the Edition model, not a
hash we can freely edit.