Re-structure DocumentExportPresenter output & export more fields
This commit adds the required information to the document export as
outlined by a comment by Kevin in this trello card -
https://trello.com/c/at9jVJVu/1036-review-contents-of-document-export-and-remove-unnecessary-fields

This changes the structure of this JSON to encompass the various models
together into a single JSON structure rather than having different parts
embedded in the hierarchy.

It applies various refactorings to the code to achieve this - mostly to
try reduce the number of methods one has to understand.
An unfortunate oddity I found was that for the various sub type classes
(e.g. NewsArticleType) the factories for these all had a different way
of initialising them which made the test slightly inconsistent.

Co-Authored-By: Kevin Dew <kevin.dew@digital.cabinet-office.gov.uk>